+++ unsloth/models/gemma2.py
@@ -125,8 +125,8 @@ def Gemma2Attention_fast_forward(
         Q, K = fast_rope_embedding(Q, K, cos, sin)
 
     if past_key_value is not None:
-        K = torch.cat([past_key_value[0], K], dim=2)
-        V = torch.cat([past_key_value[1], V], dim=2)
+        K = torch.cat([past_key_value[0], K], dim = 2)
+        V = torch.cat([past_key_value[1], V], dim = 2)
     past_key_value = (K, V) if use_cache else None
 
     # Only enable if the attention_mask is True.
@@ -152,16 +152,16 @@ def Gemma2Attention_fast_forward(
         use_varlen = seq_info is not None and past_key_value is None
 
         attention_config = AttentionConfig(
-            backend=select_attention_backend(use_varlen),
-            n_kv_heads=n_kv_heads,
-            n_groups=n_groups,
-            flash_dense_kwargs={
+            backend = select_attention_backend(use_varlen),
+            n_kv_heads = n_kv_heads,
+            n_groups = n_groups,
+            flash_dense_kwargs = {
                 "causal": True,
                 "softcap": self.config.attn_logit_softcapping,
                 "softmax_scale": self._flash_attention_softmax_scale,
                 "window_size": window,
             },
-            flash_varlen_kwargs={
+            flash_varlen_kwargs = {
                 "dropout_p": 0.0,
                 "softmax_scale": self._flash_attention_softmax_scale,
                 "causal": True,
@@ -174,20 +174,20 @@ def Gemma2Attention_fast_forward(
         # padding mask) raises. None means the byte-identical default.
         _pg_seg = resolve_prefix_seg_info(kwargs, past_key_value, attention_mask)
         context = AttentionContext(
-            bsz=bsz,
-            q_len=q_len,
-            kv_seq_len=kv_seq_len,
-            n_heads=n_heads,
-            head_dim=head_dim,
-            requires_grad=hidden_states.requires_grad,
-            seq_info=seq_info,
-            attention_mask=attention_mask,
-            causal_mask=causal_mask,
-            sliding_window=sliding_window,
-            prefix_seg_info=_pg_seg,
+            bsz = bsz,
+            q_len = q_len,
+            kv_seq_len = kv_seq_len,
+            n_heads = n_heads,
+            head_dim = head_dim,
+            requires_grad = hidden_states.requires_grad,
+            seq_info = seq_info,
+            attention_mask = attention_mask,
+            causal_mask = causal_mask,
+            sliding_window = sliding_window,
+            prefix_seg_info = _pg_seg,
         )
 
-        A = run_attention(config=attention_config, context=context, Q=Q, K=K, V=V)
+        A = run_attention(config = attention_config, context = context, Q = Q, K = K, V = V)
         A = A.reshape(bsz, q_len, n_heads * head_dim)
     else:
         fx = (
@@ -217,8 +217,8 @@ def Gemma2DecoderLayer_fast_forward(
     if use_cache and hasattr(self, "_flag_for_generation"):
         out_weight = torch.empty(
             self.input_layernorm.weight.shape,
-            dtype=torch.float32,
-            device=f"{DEVICE_TYPE_TORCH}:0",
+            dtype = torch.float32,
+            device = f"{DEVICE_TYPE_TORCH}:0",
         )
 
         residual = hidden_states
@@ -226,15 +226,15 @@ def Gemma2DecoderLayer_fast_forward(
             self.input_layernorm, hidden_states, out_weight
         )
         hidden_states, self_attn_weights, present_key_value = self.self_attn(
-            hidden_states=hidden_states,
-            causal_mask=causal_mask,
-            attention_mask=attention_mask,
-            position_ids=position_ids,
-            past_key_value=past_key_value,
-            output_attentions=output_attentions,
-            use_cache=use_cache,
-            padding_mask=padding_mask,
-            _flag_for_generation=self._flag_for_generation,
+            hidden_states = hidden_states,
+            causal_mask = causal_mask,
+            attention_mask = attention_mask,
+            position_ids = position_ids,
+            past_key_value = past_key_value,
+            output_attentions = output_attentions,
+            use_cache = use_cache,
+            padding_mask = padding_mask,
+            _flag_for_generation = self._flag_for_generation,
             **kwargs,
         )
         hidden_states = fast_rms_layernorm_inference_gemma(
@@ -253,28 +253,28 @@ def Gemma2DecoderLayer_fast_forward(
         hidden_states += residual
     else:
         residual = hidden_states
-        hidden_states = fast_rms_layernorm(self.input_layernorm, hidden_states, gemma=True)
+        hidden_states = fast_rms_layernorm(self.input_layernorm, hidden_states, gemma = True)
         hidden_states, self_attn_weights, present_key_value = self.self_attn(
-            hidden_states=hidden_states,
-            causal_mask=causal_mask,
-            attention_mask=attention_mask,
-            position_ids=position_ids,
-            past_key_value=past_key_value,
-            output_attentions=output_attentions,
-            use_cache=use_cache,
-            padding_mask=padding_mask,
+            hidden_states = hidden_states,
+            causal_mask = causal_mask,
+            attention_mask = attention_mask,
+            position_ids = position_ids,
+            past_key_value = past_key_value,
+            output_attentions = output_attentions,
+            use_cache = use_cache,
+            padding_mask = padding_mask,
             **kwargs,
         )
-        hidden_states = fast_rms_layernorm(self.post_attention_layernorm, hidden_states, gemma=True)
+        hidden_states = fast_rms_layernorm(self.post_attention_layernorm, hidden_states, gemma = True)
         hidden_states = residual + hidden_states
 
         residual = hidden_states
         hidden_states = fast_rms_layernorm(
-            self.pre_feedforward_layernorm, hidden_states, gemma=True
+            self.pre_feedforward_layernorm, hidden_states, gemma = True
         )
         hidden_states = self.mlp(hidden_states)
         hidden_states = fast_rms_layernorm(
-            self.post_feedforward_layernorm, hidden_states, gemma=True
+            self.post_feedforward_layernorm, hidden_states, gemma = True
         )
         hidden_states = residual + hidden_states
 
@@ -299,9 +299,9 @@ def Gemma2Attention_fast_forward_inference(
     hidden_states: torch.Tensor,
     past_key_value: Optional[Tuple[torch.Tensor]],
     position_ids,
-    do_prefill=False,
-    attention_mask=None,
-    use_sliding_window=False,
+    do_prefill = False,
+    attention_mask = None,
+    use_sliding_window = False,
     **kwargs,
 ):
     Xn = hidden_states
@@ -323,20 +323,20 @@ def Gemma2Attention_fast_forward_inference(
     if do_prefill:
         self.paged_attention = torch.empty(
             (KV_CACHE_INCREMENT + seq_len + 1, 2, bsz, n_kv_heads, head_dim),
-            dtype=dtype,
-            device=device,
+            dtype = dtype,
+            device = device,
         )
         self.paged_attention_K = self.paged_attention[:, 0]
         self.paged_attention_V = self.paged_attention[:, 1]
         self.paged_attention_K[:seq_len] = K1.permute(2, 0, 1, 3)
         self.paged_attention_V[:seq_len] = V1.permute(2, 0, 1, 3)
-        self.temp_QA = torch.empty((2, bsz, 1, attention_size), dtype=dtype, device=device)
-        self.temp_KV = torch.empty((2, bsz, 1, n_kv_heads * head_dim), dtype=dtype, device=device)
-        self.RH_Q = torch.empty((bsz, n_heads, 1, head_dim), dtype=dtype, device=device)
+        self.temp_QA = torch.empty((2, bsz, 1, attention_size), dtype = dtype, device = device)
+        self.temp_KV = torch.empty((2, bsz, 1, n_kv_heads * head_dim), dtype = dtype, device = device)
+        self.RH_Q = torch.empty((bsz, n_heads, 1, head_dim), dtype = dtype, device = device)
         # Only for Gemma2.
-        self.temp_O = torch.empty((bsz, 1, hidden_size), dtype=dtype, device=device)
+        self.temp_O = torch.empty((bsz, 1, hidden_size), dtype = dtype, device = device)
         self.attention = torch.empty(
-            (bsz, n_heads, 1, KV_CACHE_INCREMENT + seq_len), dtype=dtype, device=device
+            (bsz, n_heads, 1, KV_CACHE_INCREMENT + seq_len), dtype = dtype, device = device
         )
 
         # Gemma 9b should use 256, not hidden_size // num_attention_heads (224); 27b uses the derived value,
@@ -360,9 +360,9 @@ def Gemma2Attention_fast_forward_inference(
         self.paged_attention_V = self.paged_attention[:, 1]
         self.attention.resize_((bsz, n_heads, 1, self.attention.shape[-1] + KV_CACHE_INCREMENT))
 
-    Qn = fast_linear_forward(self.q_proj, Xn, out=self.temp_QA[0])
-    Kn = fast_linear_forward(self.k_proj, Xn, out=self.temp_KV[0])
-    Vn = fast_linear_forward(self.v_proj, Xn, out=self.temp_KV[1])
+    Qn = fast_linear_forward(self.q_proj, Xn, out = self.temp_QA[0])
+    Kn = fast_linear_forward(self.k_proj, Xn, out = self.temp_KV[0])
+    Vn = fast_linear_forward(self.v_proj, Xn, out = self.temp_KV[1])
     Qn = Qn.view(bsz, 1, n_heads, head_dim).transpose(1, 2)
     Kn = Kn.view(bsz, 1, n_kv_heads, head_dim).transpose(1, 2)
     Vn = Vn.view(bsz, 1, n_kv_heads, head_dim).transpose(1, 2)
@@ -415,7 +415,7 @@ def Gemma2Attention_fast_forward_inference(
     # if PyTorch adds a softcap param, consider SDPA for bsz > 1 to match llama/qwen3.
     Qn *= self.scalar
     # (Q * scalar) @ K beats (Q @ K) * scalar for stopping overflows; see ggerganov/llama.cpp#7805 (comment 2153349963).
-    A = torch_matmul(Qn, Knn.transpose(2, 3), out=self.attention[:, :, :, :cached_len])
+    A = torch_matmul(Qn, Knn.transpose(2, 3), out = self.attention[:, :, :, :cached_len])
 
     # Softcapping must happen BEFORE the mask is applied; see google-deepmind/gemma _modules.py and
     # transformers gemma2 eager_attention_forward.
@@ -429,11 +429,11 @@ def Gemma2Attention_fast_forward_inference(
             attention_mask = attention_mask[:, :, :, -A.shape[-1] :]
         A += attention_mask
 
-    A[:] = torch_nn_functional_softmax(A, dim=-1, dtype=torch.float32)
-    A = torch_matmul(A, Vnn, out=Qn)
+    A[:] = torch_nn_functional_softmax(A, dim = -1, dtype = torch.float32)
+    A = torch_matmul(A, Vnn, out = Qn)
     A = A.transpose(1, 2)
     A = A.reshape(bsz, 1, attention_size)
-    A = fast_linear_forward(self.o_proj, A, out=self.temp_O)
+    A = fast_linear_forward(self.o_proj, A, out = self.temp_O)
     return A, (Kn, Vn)
 
 
@@ -443,14 +443,14 @@ def Gemma2Model_fast_forward_inference(
     input_ids,
     past_key_values,
     position_ids,
-    attention_mask=None,
+    attention_mask = None,
     **kwargs,
 ):
     out_weights = tuple(
         torch.empty_like(
             self.model.layers[0].input_layernorm.weight,
-            dtype=torch.float32,
-            device=torch.device(x),
+            dtype = torch.float32,
+            device = torch.device(x),
         )
         for x in range(DEVICE_COUNT)
     )
@@ -458,7 +458,7 @@ def Gemma2Model_fast_forward_inference(
     hidden_states = self.model.embed_tokens(input_ids)
     hidden_states = hidden_states.to(_get_dtype(dtype_from_config(self.config)))
     # 3072**0.5 is 55.5000 in bfloat16 against 55.4256 in float32, and 2048**0.5 is 45.2500 against 45.2548.
-    hidden_states *= torch.tensor(math_sqrt(self.config.hidden_size), dtype=hidden_states.dtype)
+    hidden_states *= torch.tensor(math_sqrt(self.config.hidden_size), dtype = hidden_states.dtype)
 
     bsz, q_len, hd = hidden_states.shape
     seq_len = past_key_values[0][0].shape[-2]
@@ -472,7 +472,7 @@ def Gemma2Model_fast_forward_inference(
                 (bsz, q_len),
                 hidden_states,
                 seq_len,
-                sliding_window=self.config.sliding_window,
+                sliding_window = self.config.sliding_window,
             )
             GA = _prepare_4d_causal_attention_mask_for_sdpa(
                 attention_mask,
@@ -497,12 +497,12 @@ def Gemma2Model_fast_forward_inference(
         )
         hidden_states, present_key_value = Gemma2Attention_fast_forward_inference(
             decoder_layer.self_attn,
-            hidden_states=hidden_states,
-            past_key_value=past_key_values[idx],
-            position_ids=position_ids,
-            attention_mask=SWA if use_sliding_window else GA,
-            do_prefill=not hasattr(decoder_layer.self_attn, "paged_attention"),
-            use_sliding_window=use_sliding_window,
+            hidden_states = hidden_states,
+            past_key_value = past_key_values[idx],
+            position_ids = position_ids,
+            attention_mask = SWA if use_sliding_window else GA,
+            do_prefill = not hasattr(decoder_layer.self_attn, "paged_attention"),
+            use_sliding_window = use_sliding_window,
         )
         hidden_states = fast_rms_layernorm_inference_gemma(
             decoder_layer.post_attention_layernorm,
@@ -531,10 +531,10 @@ def Gemma2Model_fast_forward_inference(
     )
 
     return BaseModelOutputWithPast(
-        last_hidden_state=hidden_states,
-        past_key_values=next_decoder_cache,
-        hidden_states=[],
-        attentions=[],
+        last_hidden_state = hidden_states,
+        past_key_values = next_decoder_cache,
+        hidden_states = [],
+        attentions = [],
     )
 
 
@@ -542,10 +542,10 @@ class FastGemma2Model(FastLlamaModel):
     @staticmethod
     def pre_patch():
         init_name, function = patch_linear_scaling(
-            model_name="gemma2",
-            rope_module=GemmaFixedRotaryEmbedding,
-            scaled_rope_module=GemmaFixedLinearScalingRotaryEmbedding,
-            attention_module=Gemma2Attention,
+            model_name = "gemma2",
+            rope_module = GemmaFixedRotaryEmbedding,
+            scaled_rope_module = GemmaFixedLinearScalingRotaryEmbedding,
+            attention_module = Gemma2Attention,
         )
         if init_name is not None:
             exec(function, globals())
@@ -570,11 +570,11 @@ class FastGemma2Model(FastLlamaModel):
     def post_patch(
         model,
         tokenizer,
-        correct_dtype=None,
+        correct_dtype = None,
     ):
         # Gemma does not downcast RoPE.
         model, tokenizer = patch_model_and_tokenizer(
-            model, tokenizer, downcast_rope=False, correct_dtype=correct_dtype
+            model, tokenizer, downcast_rope = False, correct_dtype = correct_dtype
         )
 
         # Gemma returns output * (1 + self.weight); see transformers models/gemma/modeling_gemma.py#L89.
